refactor: api 모듈화, 불필요한 파일 제거
api router에서 콜백 함수로 api 모듈 호출하는 방식으로 변경
