+++ server/routes/apis/columnsApi.js
@@ -1,0 +1,3 @@
+export default (req, res) => {
+  res.json({ result: [{ column_id: 1, title: "column" }] });
+};
