+++ server/routes/index.js
@@ -1,9 +1,11 @@
-import express from "express";
-const router = express.Router();
+import { Router } from "express";
+import todosApi from "./apis/todosApi";
+import columnsApi from "./apis/columnsApi";
+
+const router = Router();
 
 /* GET home page. */
-router.get("/", function (req, res, next) {
-  res.json({ title: "Express" });
-});
+router.get("/todos", todosApi);
+router.get("/columns", columnsApi);
 
 module.exports = router;
